pyproject.toml
[project]
name = "odoo-mcp"
version = "1.0.0"
description = "Standalone FastMCP server for Odoo purchasing automation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.1.1",
    "pydantic-settings>=2.13.1",
    "pytesseract>=0.3.10",
    "opencv-python-headless>=4.9",
    "pillow>=10.0",
    "pdf2image>=1.17",
    "rapidfuzz>=3.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools", "utils"]
